requirements.txt
# requirements.txt ACTUALIZADO
asgiref==3.9.2
Django==4.2
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
gunicorn==23.0.0
dj-database-url==2.1.0
python-decouple==3.8
django-cors-headers==4.5.0
whitenoise==6.6.0
# psycopg2-binary no soporta bien Python 3.13 (termina compilando y falla).
# Usamos psycopg v3 + binarios precompilados (explicito para evitar conflictos del resolver).
psycopg==3.3.3
psycopg-binary==3.3.3
djangorestframework-simplejwt==5.3.1
